feat(telemetry): expand ring-buffer API and strengthen recorder tests
Concurrent agent improvements to the telemetry subsystem:

- disruptor.rs: minor API refinements to the ring-buffer claim/write
  interface for ergonomics
- recorder.rs: recorder integration adjustments matching the updated
  disruptor API
- tests/recorder_nomock.rs: substantially expanded no-mock integration
  tests with additional coverage for flush timing, shutdown ordering,
  and parquet persistence validation
- .beads/issues.jsonl: concurrent issue tracker updates

+++ crates/pt-telemetry/src/disruptor.rs
@@ -276,7 +276,12 @@ mod tests {
                     after_claim.wait();
                     rb.write_event(
                         seq,
-                        FixedSizeEvent::from_parts(thread_id as u64, thread_id, thread_id, "x"),
+                        FixedSizeEvent::from_parts(
+                            thread_id as u64,
+                            thread_id as u32,
+                            thread_id as u32,
+                            "x",
+                        ),
                     );
                     rb.commit(seq);
                     seq
